# dymer-webserver/swagger_webserver.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "DYMER API",
    "description": "Documentation of <b>Webserver</b> module."
  },
  "host": "localhost:0000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/templates/api/v1/template/": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/content/{fileid}": {
      "get": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "produces": [
          null
        ],
        "parameters": [
          {
            "name": "fileid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/api/templates/api/v1/template/create": {
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/addAsset": {
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/update": {
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/updateAsset": {
      "post": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "req",
            "in": "query",
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/{id}": {
      "delete": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/templates/api/v1/template/{id}/{fid}": {
      "delete": {
        "tags": [
          "Templates"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "fid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "dymeruser",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}